feat(as/token): extract shared generic signer for token brokers
Extract the duplicated signer trait and signer implementations from the
simple, oidc, and ear token brokers into one shared generic module
(`attestation-service/src/token/signer.rs`).

- `SignKeyProvider<K>`: a generic, object-safe trait (private_key,
  cert_chain, cert_url, cert_pem_raw). Brokers hold
  `Arc<dyn SignKeyProvider<RsaPrivateKey>>` (simple/oidc) or
  `Arc<dyn SignKeyProvider<SecretKey>>` (ear).
- `FsSigner<K>` and `EphemeralSigner<K>`: two generic structs; the shared
  cert-chain/cert-url/cert-pem-raw plumbing is written once in the generic
  trait impls.
- K-specific construction ("specialization") via concrete inherent impls:
  `FsSigner::<RsaPrivateKey>::from_config` (PKCS#8 + PKCS#1 fallback),
  `FsSigner::<SecretKey>::from_config` (SEC1 + PKCS#8 fallback),
  `EphemeralSigner::<RsaPrivateKey>::new()` (RSA 2048),
  `EphemeralSigner::<SecretKey>::new()` (EC P-256).
- Shared `SignerConfig` DTO, re-exported from each broker so all existing
  paths (`simple::SignerConfig`, etc.) keep working.

JWT assembly, JWK generation, alg/hash constants, and per-broker
`from_config` dispatch stay in the brokers — behavior is byte-identical
(verified by the cert-chain x5c tests and the EC key round-trip test).

The `FsSigner` trait impl is `#[cfg(feature = "fs")]`-gated (mirroring the
original `ConfigSigner` gating): its `.context()` on
`Result<_, rustls_pki_types::pem::Error>` requires `pem::Error:
std::error::Error`, a bound not satisfied on wasm32-unknown-unknown. The
trait def, `CertificateDer` import, and the `EphemeralSigner` impl stay
un-gated, preserving the no-fs ephemeral-signing path.
